refactor: improve modularity of result operations
