+++ amira/pre_process_operations.py
@@ -70,6 +70,7 @@ def determine_gene_strand(read: pysam.libcalignedsegment.AlignedSegment) -> tupl
         gene_name = "+" + strandlessGene
     return gene_name, strandlessGene
 
+
 def remove_poorly_mapped_genes(
     pandora_consensus,
     zero_coverage_threshold,
@@ -219,3 +220,23 @@ def convert_pandora_output(
                 subsettedGenesOfInterest.add(annotatedReads[r][g][1:])
     assert not len(annotatedReads) == 0
     return annotatedReads, subsettedGenesOfInterest, gene_position_dict
+
+
+def process_reference_alleles(path_to_interesting_genes):
+    # import the list of genes of interest
+    with open(path_to_interesting_genes, "r") as i:
+        reference_content = i.read().split(">")[1:]
+    genesOfInterest = set()
+    reference_alleles = {}
+    for allele in reference_content:
+        newline_split = allele.split("\n")
+        assert (
+            newline_split[0].count(";") == 1
+        ), "Reference FASTA headers can only contain 1 semicolon"
+        gene_name, allele_name = newline_split[0].split(";")
+        genesOfInterest.add(gene_name)
+        sequence = "".join(newline_split[1:])
+        if gene_name not in reference_alleles:
+            reference_alleles[gene_name] = {}
+        reference_alleles[gene_name][allele_name] = sequence
+    return reference_alleles, genesOfInterest
